+++ drake/systems/plants/RigidBody.cpp
@@ -21,7 +21,7 @@ RigidBody::RigidBody()
   model_id_ = 0;
   position_num_start = 0;
   velocity_num_start = 0;
-  body_index = 0;
+  body_index_ = 0;
   mass = 0.0;
   com = Vector3d::Zero();
   I << drake::SquareTwistMatrix<double>::Zero();
@@ -60,6 +60,10 @@ const RigidBody* RigidBody::get_parent() const { return parent_; }
 RigidBody* RigidBody::get_mutable_parent() { return parent_; }
 
 bool RigidBody::hasParent() const { return parent_ != nullptr; }
+
+void RigidBody::set_body_index(int body_index) { body_index_ = body_index; }
+
+int RigidBody::get_body_index() const { return body_index_; }
 
 void RigidBody::addVisualElement(const DrakeShapes::VisualElement& element) {
   visual_elements.push_back(element);