+++ drake/systems/plants/parser_sdf.cc
@@ -310,7 +310,7 @@ bool parseSDFLink(RigidBodyTree* model, std::string model_name,
   }
 
   model->add_rigid_body(std::move(owned_body));
-  *index = body->body_index;
+  *index = body->get_body_index();
   return true;
 }