+++ drake/systems/plants/test/rigid_body_tree/rigid_body_tree_test.cc
@@ -66,7 +66,7 @@ TEST_F(RigidBodyTreeTest, TestAddFloatingJointNoOffset) {
   // Adds floating joints that connect r1b1 and r2b1 to the rigid body tree's
   // world at zero offset.
   tree->AddFloatingJoint(DrakeJoint::QUATERNION,
-                         {r1b1->body_index, r2b1->body_index});
+                         {r1b1->get_body_index(), r2b1->get_body_index()});
 
   // Verfies that the two rigid bodies are located in the correct place.
   const DrakeJoint& jointR1B1 = tree->FindBody("body1", "robot1")->getJoint();
@@ -102,7 +102,8 @@ TEST_F(RigidBodyTreeTest, TestAddFloatingJointWithOffset) {
       T_r1and2_to_world);
 
   tree->AddFloatingJoint(DrakeJoint::QUATERNION,
-                         {r1b1->body_index, r2b1->body_index}, weld_to_frame);
+                         {r1b1->get_body_index(), r2b1->get_body_index()},
+                         weld_to_frame);
 
   // Verfies that the two rigid bodies are located in the correct place.
   const DrakeJoint& jointR1B1 = tree->FindBody("body1", "robot1")->getJoint();
@@ -123,7 +124,7 @@ TEST_F(RigidBodyTreeTest, TestAddFloatingJointWeldToLink) {
   // unique_ptr_reference's
   tree->add_rigid_body(std::unique_ptr<RigidBody>(r1b1));
 
-  tree->AddFloatingJoint(DrakeJoint::QUATERNION, {r1b1->body_index});
+  tree->AddFloatingJoint(DrakeJoint::QUATERNION, {r1b1->get_body_index()});
 
   // Adds rigid body r2b1 to the rigid body tree and welds it to r1b1 with
   // offset x = 1, y = 1, z = 1. Verifies that it is in the correct place.
@@ -143,7 +144,8 @@ TEST_F(RigidBodyTreeTest, TestAddFloatingJointWeldToLink) {
       Eigen::aligned_allocator<RigidBodyFrame>(), "body1",
       tree->FindBody("body1", "robot1"), T_r2_to_r1);
 
-  tree->AddFloatingJoint(DrakeJoint::QUATERNION, {r2b1->body_index}, r2b1_weld);
+  tree->AddFloatingJoint(DrakeJoint::QUATERNION, {r2b1->get_body_index()},
+      r2b1_weld);
 
   // Adds rigid body r3b1 and r4b1 to the rigid body tree and welds it to r2b1
   // with offset x = 2, y = 2, z = 2. Verifies that it is in the correct place.
@@ -166,7 +168,7 @@ TEST_F(RigidBodyTreeTest, TestAddFloatingJointWeldToLink) {
       tree->FindBody("body1", "robot2"), T_r3_and_r4_to_r2);
 
   tree->AddFloatingJoint(DrakeJoint::QUATERNION,
-                         {r3b1->body_index, r4b1->body_index},
+                         {r3b1->get_body_index(), r4b1->get_body_index()},
                          r3b1_and_r4b1_weld);
 
   EXPECT_TRUE(tree->FindBody("body1", "robot1")